add tests for previous fixes and mocking tests should only run once and not for each driver
